getopts and options for builds

+++ build_and_notify_single_project.sh
@@ -1,13 +1,31 @@
 #!/bin/sh
 
 my_dir="`dirname \"$0\"`"
-cd 
 if [ ! -d "$my_dir" ] ; then
 	echo "Could not find directory $my_dir" >&2
 	exit 5
 fi
 
 . $my_dir/functions.sh
+
+usage() {
+	echo "Usage: $0 [-c xcodebuild configuration] [-m mobileprovision file] [-p xcodeproj file ] [-t build target] <Project Name>" >&2
+	echo "	-c xcodebuild configuration: Default is 'Debug'" >&2
+	echo "	-m mobileprovision file: Default is most recent Team Provisioning Profile in $HOME/Library/MobileDevice/Provisioning Profiles" >&2
+	echo "	-p xcodeproj file: Default is <Project Name>.xcodeproj" >&2
+	echo "	-t built target: Default is first target found in the output of 'xcodebuild -list'" >&2
+	exit 3
+}
+
+while getopts "c:p:t:m:" optionName; do
+	case "$optionName" in
+		c) configuration="$OPTARG"; shift; shift;;
+		m) provprofile="$OPTARG"; shift; shift;;
+		p) projectFile="$OPTARG"; shift; shift;;
+		t) build_target="$OPTARG"; shift; shift;;
+		h) usage;;
+	esac
+done
 
 if [ -z "$SCM_WORKING_DIR" -o ! -d "$SCM_WORKING_DIR" ] ; then
 	echo "Undefined or incorrect SCM_WORKING_DIR variable.  Please check your XCAB.settings file or environment" >&2
@@ -21,25 +39,41 @@ fi
 
 #Need to get current project directory name and URL from the command line
 if [ $# -ne 1 ] ; then
-	echo "Usage: $0 <Project Name>" >&2
-	exit 3
+	echo "Could not get "
+	usage
 fi
 
 target="$1"
+
+#Setup Defaults
+if [ -z "$configuration" ] ; then
+	configuration="Debug"
+fi
+if [ -z "$projectFile" ] ; then
+	projectFile="${target}.xcodeproj"
+fi
+
+if [ -z "$provprofile" ] ; then
+	#Find the most recent automatically generated provisioning profile
+	for f in `ls -1tr "$HOME/Library/MobileDevice/Provisioning Profiles/"`; do 
+		grep -l 'Team Provisioning Profile: *' "$HOME/Library/MobileDevice/Provisioning Profiles/$f" > /dev/null 2>&1
+		if [ $? -eq 0 ]; then
+			provprofile="$HOME/Library/MobileDevice/Provisioning Profiles/$f"
+		fi
+	done
+fi
+
+if [ -z "$provprofile" ] ; then
+	echo "Error: Could not find a provisioning profile. Please consider specifying one with the -m option to $0" >&2
+	usage
+	exit 3
+fi
 
 if [ ! -d "$SCM_WORKING_DIR/$target" ] ; then
 	echo "Error: Could not find directory for project in $SCM_WORKING_DIR/$target" >&2
 	echo "Usage: $0 <Project Name>" >&2
 	exit 3
 fi
-
-#Find the most recent automatically generated provisioning profile
-for f in `ls -1tr "$HOME/Library/MobileDevice/Provisioning Profiles/"`; do 
-	grep -l 'Team Provisioning Profile: *' "$HOME/Library/MobileDevice/Provisioning Profiles/$f" > /dev/null 2>&1
-	if [ $? -eq 0 ]; then
-		provprofile="$HOME/Library/MobileDevice/Provisioning Profiles/$f"
-	fi
-done
 
 cd "$SCM_WORKING_DIR/$target"
 
@@ -48,11 +82,6 @@ cd "$SCM_WORKING_DIR/$target"
 #TODO make sure that, if there is an error, it's only
 #  a connection error before we ignore it
 git fetch > /dev/null 2>&1
-
-if [ x"`ls -1d *xcodeproj 2>/dev/null`" == x ] ; then
-	#Not an iphone dir
-	continue
-fi
 
 build_time_human="`date +%Y%m%d%H%M%S`"
 
@@ -63,10 +92,10 @@ cutoff_time="`expr $now - $cutoff_window`"
 
 already_built="`cat $OVER_AIR_INSTALLS_DIR/$target/*/sha.txt 2>/dev/null`"
 
-for candidate in `git branch -a | sed -e 's/^..//'` ; do
+for candidate in `git branch -a | sed -e 's/^..//' -e 's/ ->.*$//' -e 's,^remotes/,,' | sort -t / -k 2 -k 1 -k 3` ; do
 	sha="`git rev-parse $candidate`"
 	commit_time="`git log -1 --pretty=format:"%ct" $sha`"
-	if [ $commit_time -gt $cutoff_time ] ; then
+	if [ "$commit_time" -gt "$cutoff_time" ] ; then
 		is_built="`echo $already_built | grep $sha`"
 		if [ -z "$is_built" ] ; then
 			if [ ! -z "`echo "$candidate" | grep '/'`" ] ; then
@@ -83,16 +112,20 @@ for candidate in `git branch -a | sed -e 's/^..//'` ; do
 			git clean -d -f -x
 			mkdir -p "$OVER_AIR_INSTALLS_DIR/$target/$build_time_human/"
 			mkdir -p "$OVER_AIR_INSTALLS_DIR/$target/tmp/$build_time_human/"
-			#TODO need to figure out a way to indicate that the user wants to build other targets
-			build_target=`xcodebuild -list | awk '$1=="Targets:",$1==""' | grep -v "Targets:" | grep -v "^$" | sed -e 's/^  *//' | head -1`
+			if [ -z "$build_target" ] ; then
+				build_target=`xcodebuild -list -project $projectFile | awk '$1=="Targets:",$1==""' | grep -v "Targets:" | grep -v "^$" | sed -e 's/^  *//' | head -1`
+			fi
+			
 			#TODO need to make sure we're building for the device
-			xcodebuild build -target $build_target -configuration Debug > "$OVER_AIR_INSTALLS_DIR/$target/$build_time_human/${build_target}_output.txt" 2> "$OVER_AIR_INSTALLS_DIR/$target/$build_time_human/${build_target}_error.txt"
+			xcodebuild build -target $build_target -configuration $configuration -project $projectFile > "$OVER_AIR_INSTALLS_DIR/$target/$build_time_human/${build_target}_output.txt" 2> "$OVER_AIR_INSTALLS_DIR/$target/$build_time_human/${build_target}_error.txt"
 			if [ $? -ne 0 ] ; then
 				echo "Build Failed" >&2
 				echo "$sha" > "$OVER_AIR_INSTALLS_DIR/$target/$build_time_human/sha.txt" #Don't try to build this again - it would fail over and over
 				rm -rf "$OVER_AIR_INSTALLS_DIR/$target/tmp/$build_time_human/"
 				exit 3
 			fi
+			#Don't build this again this run if more than one branch points at same sha
+			already_built="$already_built $sha"
 			
 			App_location="`grep ${build_target}.app $OVER_AIR_INSTALLS_DIR/$target/$build_time_human/${build_target}_output.txt | grep '^CodeSign' | sed -e 's/^CodeSign //'`"
 			
@@ -107,15 +140,17 @@ for candidate in `git branch -a | sed -e 's/^..//'` ; do
 				rm -rf "$OVER_AIR_INSTALLS_DIR/$target/tmp/$build_time_human/"
 				exit 3
 			fi
-			
-			"/Applications/BetaBuilder for iOS Apps.app/Contents/MacOS/BetaBuilder for iOS Apps" -ipaPath="$OVER_AIR_INSTALLS_DIR/$target/tmp/$build_time_human/${build_target}.ipa" -outputDirectory="$OVER_AIR_INSTALLS_DIR/$target/$build_time_human" -webserver="${XCAB_WEB_ROOT}/${target}/$build_time_human"
-			if [ $? -ne 0 ] ; then
-				rm -rf $OVER_AIR_INSTALLS_DIR/$target/tmp/$build_time_human/${build_target}.ipa
-				echo "BetaBuilder for iOS Apps Failed" >&2
-				echo "$sha" > "$OVER_AIR_INSTALLS_DIR/$target/$build_time_human/sha.txt" #Don't try to build this again - it would fail over and over
-				rm -rf "$OVER_AIR_INSTALLS_DIR/$target/tmp/$build_time_human/"
-				exit 3
-			else
+						
+			if [ ! -z "${XCAB_WEB_ROOT}" -a -f "/Applications/BetaBuilder for iOS Apps.app/Contents/MacOS/BetaBuilder for iOS Apps" ] ; then
+				"/Applications/BetaBuilder for iOS Apps.app/Contents/MacOS/BetaBuilder for iOS Apps" -ipaPath="$OVER_AIR_INSTALLS_DIR/$target/tmp/$build_time_human/${build_target}.ipa" -outputDirectory="$OVER_AIR_INSTALLS_DIR/$target/$build_time_human" -webserver="${XCAB_WEB_ROOT}/${target}/$build_time_human"
+				if [ $? -ne 0 ] ; then
+					rm -rf $OVER_AIR_INSTALLS_DIR/$target/tmp/$build_time_human/${build_target}.ipa
+					echo "BetaBuilder for iOS Apps Failed" >&2
+					echo "$sha" > "$OVER_AIR_INSTALLS_DIR/$target/$build_time_human/sha.txt" #Don't try to build this again - it would fail over and over
+					rm -rf "$OVER_AIR_INSTALLS_DIR/$target/tmp/$build_time_human/"
+					exit 3
+				fi
+				
 				#Save off the symbols, too
 				if [ -d "${App_location}.dSYM" ] ; then
 					tar czf "$OVER_AIR_INSTALLS_DIR/$target/$build_time_human/${build_target}.app.dSYM.tar.gz" "${App_location}.dSYM"
@@ -127,18 +162,19 @@ for candidate in `git branch -a | sed -e 's/^..//'` ; do
 					else
 						TESTFLIGHT_NOTIFY="-F notify=False"
 					fi
-					curl http://testflightapp.com/api/builds.json -F file=@"$OVER_AIR_INSTALLS_DIR/$target/tmp/$build_time_human/${build_target}.ipa"  -F api_token="$TESTFLIGHT_API_TOKEN" -F team_token="$TESTFLIGHT_TEAM" -F notes='New ${target} Build was uploaded via the upload API' $TESTFLIGHT_NOTIFY 
+					curl http://testflightapp.com/api/builds.json -F file=@"$OVER_AIR_INSTALLS_DIR/$target/$build_time_human/${build_target}.ipa"  -F api_token="$TESTFLIGHT_API_TOKEN" -F team_token="$TESTFLIGHT_TEAM" -F notes='New ${target} Build was uploaded via the upload API' $TESTFLIGHT_NOTIFY 
 				fi
 				
+				#Clean up temp dir
 				rm -rf "$OVER_AIR_INSTALLS_DIR/$target/tmp/$build_time_human/${build_target}.ipa"
-				
+		
 				if [ ! -z "$RSYNC_USER" ] ; then
 					#If we're not using Dropbox's public web server, run rsync now
 					rsync -r ${OVER_AIR_INSTALLS_DIR} ${RSYNC_USER}@${XCAB_WEBSERVER_HOSTNAME}:${XCAB_WEBSERVER_XCAB_DIRECTORY_PATH}
 				fi
-			
+		
 				wait_for_idle_dropbox
-				
+			
 				#Wait to make sure the file has appeared on the server
 				IPA_STATUS="`curl -sI -X HEAD \"${XCAB_WEB_ROOT}/${target}/$build_time_human/${build_target}.ipa\" | grep HTTP/1.1 | awk '{print $2}'`"
 				PLIST_STATUS="`curl -sI -X HEAD \"${XCAB_WEB_ROOT}/${target}/$build_time_human/manifest.plist\" | grep HTTP/1.1 | awk '{print $2}'`"
@@ -157,18 +193,38 @@ for candidate in `git branch -a | sed -e 's/^..//'` ; do
 						exit 4
 					fi
 				done
-				
+			
 				#We're making the implicit assumption here that there aren't 
 				#  going to be a bunch of new changes per run
 				#   so it won't spam the user to notify for each one
 				#Notify with Boxcar
 				$my_dir/notify_with_boxcar.sh "notification[source_url]=${XCAB_WEB_ROOT}/$target/$build_time_human/index.html" "notification[message]=New+${target}+Build+available"
+			else
+				#Put in the permenant location
+				cp "$OVER_AIR_INSTALLS_DIR/$target/tmp/$build_time_human/${build_target}.ipa" "$OVER_AIR_INSTALLS_DIR/$target/$build_time_human/${build_target}.ipa"
+				
+				#Save off the symbols, too
+				if [ -d "${App_location}.dSYM" ] ; then
+					tar czf "$OVER_AIR_INSTALLS_DIR/$target/$build_time_human/${build_target}.app.dSYM.tar.gz" "${App_location}.dSYM"
+				fi
+				
+				if [ ! -z "$TESTFLIGHT_API_TOKEN" -a ! -z "$TESTFLIGHT_TEAM" ] ; then
+					if [ ! -z "$TESTFLIGHT_DIST" ] ; then
+						TESTFLIGHT_NOTIFY="-F notify=True  -F distribution_lists=\"$TESTFLIGHT_DIST\""
+					else
+						TESTFLIGHT_NOTIFY="-F notify=False"
+					fi
+					curl http://testflightapp.com/api/builds.json -F file=@"$OVER_AIR_INSTALLS_DIR/$target/$build_time_human/${build_target}.ipa"  -F api_token="$TESTFLIGHT_API_TOKEN" -F team_token="$TESTFLIGHT_TEAM" -F notes='New ${target} Build was uploaded via the upload API' $TESTFLIGHT_NOTIFY 
+				else
+					echo "ipa saved to $OVER_AIR_INSTALLS_DIR/$target/$build_time_human/" >&2
+					echo "but TestFlight credentials not defined and Beta Builder not installed (or web root not configured)" >&2
+					echo "So it's not very useful, except maybe as a backup for your app builds.  Consider adding one of those services" >&2
+				fi
+				
 			fi
-								
+
 			echo "$sha" > "$OVER_AIR_INSTALLS_DIR/$target/$build_time_human/sha.txt"
 			rm -rf "$OVER_AIR_INSTALLS_DIR/$target/tmp/$build_time_human/"
 		fi
 	fi
-	#Don't build this again if more than one branch points at same sha
-	already_built="$already_built $sha"
 done
